# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Civil Construction ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vendors:
    get:
      operationId: listVendors
      tags: [vendors]
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vendor"
    post:
      operationId: createVendor
      tags: [vendors]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVendorBody"
      responses:
        "201":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /vendors/{id}:
    get:
      operationId: getVendor
      tags: [vendors]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
    patch:
      operationId: updateVendor
      tags: [vendors]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVendorBody"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"

  # Inventory
...
  /vehicles:
    get:
      operationId: listVehicles
      tags: [vehicles]
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [available, in_use, maintenance, out_of_service]
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vehicle"
    post:
      operationId: createVehicle
      tags: [vehicles]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVehicleBody"
      responses:
        "201":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
  /vehicles/{id}:
    get:
      operationId: getVehicle
      tags: [vehicles]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
    patch:
      operationId: updateVehicle
      tags: [vehicles]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVehicleBody"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
  /vehicles/maintenance:
    get:
      operationId: listMaintenanceRecords
      tags: [vehicles]
      parameters:
        - name: vehicleId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MaintenanceRecord"
    post:
      operationId: createMaintenanceRecord
      tags: [vehicles]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMaintenanceRecordBody"
      responses:
        "201":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceRecord"
  /vehicles/fuel-logs:
    get:
      operationId: listFuelLogs
      tags: [vehicles]
      parameters:
        - name: vehicleId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FuelLog"
    post:
      operationId: createFuelLog
      tags: [vehicles]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFuelLogBody"
      responses:
        "201":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FuelLog"
...
components:
  schemas:
...
    Vendor:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        contactPerson:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        gstNumber:
          type: string
        panNumber:
          type: string
        category:
          type: string
        rating:
          type: number
        status:
          type: string
          enum: [active, inactive, blacklisted]
        createdAt:
          type: string
      required: [id, name, phone, category, rating, status, createdAt]
    CreateVendorBody:
      type: object
      properties:
        name:
          type: string
        contactPerson:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        gstNumber:
          type: string
        panNumber:
          type: string
        category:
          type: string
      required: [name, phone, category]
    UpdateVendorBody:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
        rating:
          type: number
        contactPerson:
          type: string
        email:
          type: string
        phone:
          type: string
...
    Vehicle:
      type: object
      properties:
        id:
          type: integer
        registrationNumber:
          type: string
        make:
          type: string
        model:
          type: string
        type:
          type: string
          enum: [truck, excavator, crane, mixer, tipper, JCB, other]
        year:
          type: integer
        status:
          type: string
          enum: [available, in_use, maintenance, out_of_service]
        driver:
          type: string
        assignedProject:
          type: string
        lastService:
          type: string
        nextService:
          type: string
        fuelType:
          type: string
        odometer:
          type: number
        insuranceExpiry:
          type: string
        createdAt:
          type: string
      required: [id, registrationNumber, make, model, type, year, status, createdAt]
    CreateVehicleBody:
      type: object
      properties:
        registrationNumber:
          type: string
        make:
          type: string
        model:
          type: string
        type:
          type: string
        year:
          type: integer
        fuelType:
          type: string
        driver:
          type: string
        insuranceExpiry:
          type: string
      required: [registrationNumber, make, model, type, year, fuelType]
    UpdateVehicleBody:
      type: object
      properties:
        status:
          type: string
        driver:
          type: string
        assignedProject:
          type: string
        odometer:
          type: number
        nextService:
          type: string
    MaintenanceRecord:
      type: object
      properties:
        id:
          type: integer
        vehicleId:
          type: integer
        vehicleRegistration:
          type: string
        type:
          type: string
          enum: [scheduled, breakdown, inspection]
        description:
          type: string
        cost:
          type: number
        date:
          type: string
        completedDate:
          type: string
        status:
          type: string
          enum: [pending, in_progress, completed]
        mechanicName:
          type: string
        createdAt:
          type: string
      required: [id, vehicleId, vehicleRegistration, type, description, cost, date, status, createdAt]
    CreateMaintenanceRecordBody:
      type: object
      properties:
        vehicleId:
          type: integer
        type:
          type: string
          enum: [scheduled, breakdown, inspection]
        description:
          type: string
        cost:
          type: number
        date:
          type: string
        mechanicName:
          type: string
      required: [vehicleId, type, description, cost, date]
    FuelLog:
      type: object
      properties:
        id:
          type: integer
        vehicleId:
          type: integer
        vehicleRegistration:
          type: string
        liters:
          type: number
        costPerLiter:
          type: number
        totalCost:
          type: number
        odometer:
          type: number
        date:
          type: string
        filledBy:
          type: string
        createdAt:
          type: string
      required: [id, vehicleId, vehicleRegistration, liters, costPerLiter, totalCost, date, createdAt]
    CreateFuelLogBody:
      type: object
      properties:
        vehicleId:
          type: integer
        liters:
          type: number
        costPerLiter:
          type: number
        odometer:
          type: number
        date:
          type: string
        filledBy:
          type: string
      required: [vehicleId, liters, costPerLiter, odometer, date]
...